+++ R/VERBS-transmute.R
@@ -11,15 +11,18 @@ dplyr::transmute
 #' @name transmute-table.express
 #' @export
 #' @importFrom rlang expr
+#' @importFrom rlang warn
 #'
 #' @template data-arg
 #' @param ... Clauses for transmuting columns. For `j` inside the `data.table`'s frame.
+#' @param .enlist See details.
 #' @template parse-arg
 #' @template chain-arg
 #'
 #' @details
 #'
-#' Everything in `...` is wrapped in a call to `list`.
+#' Everything in `...` is wrapped in a call to `list` by default. If only one expression is given,
+#' you can set `.enlist` to `FALSE` to skip the call to `list`.
 #'
 #' @template docu-examples
 #'
@@ -30,14 +33,23 @@ dplyr::transmute
 #' data.table::as.data.table(mtcars) %>%
 #'     transmute(ans = mpg * 2)
 #'
-transmute.ExprBuilder <- function(.data, ...,
+transmute.ExprBuilder <- function(.data, ..., .enlist = TRUE,
                                   .parse = getOption("table.express.parse", FALSE),
                                   .chain = getOption("table.express.chain", TRUE))
 {
     clauses <- parse_dots(.parse, ...)
     if (length(clauses) == 0L) return(.data)
 
-    .data$set_j(rlang::expr(list(!!!clauses)), .chain)
+    if (.enlist) {
+        .data$set_j(rlang::expr(list(!!!clauses)), .chain)
+    }
+    else {
+        if (length(clauses) > 1L) {
+            rlang::warn("Only 1 expression is allowed for .enlist = FALSE, ignoring all but first.")
+        }
+
+        .data$set_j(clauses[[1L]], .chain)
+    }
 }
 
 #' @rdname transmute-table.express